backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scriptorium"
version = "0.1.0"
description = "Self-hosted book and comics library server"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Scriptorium Contributors"}
]
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0.0",
    "watchfiles>=0.21.0",
    "httpx>=0.25.0",
    "Pillow>=10.0.0",
    # Text extraction (epub2md integration)
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "pypdf>=3.17.0",
    "pdfplumber>=0.10.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
